Revert "Merged PR 120: ONNX export"
This reverts commit b0a67143468971f9cf5d732aa968f315d63fbbb5, reversing
changes made to bd23925a0254db18001816bf2e0b0a6d5086e5c5.

+++ maskrcnn_benchmark/csrc/cpu/vision.h
@@ -1,6 +1,6 @@
 // Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved.
 #pragma once
-#include <torch/script.h>
+#include <torch/extension.h>
 
 
 at::Tensor ROIAlign_forward_cpu(const at::Tensor& input,
@@ -13,4 +13,4 @@ at::Tensor ROIAlign_forward_cpu(const at::Tensor& input,
 
 at::Tensor nms_cpu(const at::Tensor& dets,
                    const at::Tensor& scores,
-                   const double threshold);
+                   const float threshold);
